Create README.md
Getting and Cleaning Data: wearable computer data project

GC_Data contains R code and documentation files for the Getting and Cleaning Data course project, This course is run by cousera and is part of the Data Science Specilization coursse.
Files uploaded includes
CodeBook.md. This is a markdown file which describes all the variables, source of data, and any processing performed to clean the data.
run_analysis.R. This is an R script file which contains all the codes written to clean or tidy the data and to write a new tidy.txt file. 

The output of the 5th step is called secTidySet.txt, and uploaded in the course project's form.